complete artefact and edges implementation
