--- bitloops_cli/src/commands/devql/support.rs
@@ -1,3 +1,0 @@
-include!("ingestion.rs");
-include!("query_pipeline.rs");
-include!("db_utils.rs");
