+++ bitloops_cli/src/engine/devql/ingestion/types.rs
@@ -1,0 +1,67 @@
+// Shared types used across ingestion modules.
+
+#[derive(Debug, Clone, Default)]
+struct IngestionCounters {
+    checkpoints_processed: usize,
+    events_inserted: usize,
+    artefacts_upserted: usize,
+    checkpoints_without_commit: usize,
+}
+
+#[derive(Debug, Clone)]
+struct CheckpointCommitInfo {
+    commit_sha: String,
+    commit_unix: i64,
+    author_name: String,
+    author_email: String,
+    subject: String,
+}
+
+#[derive(Debug, Clone)]
+struct FileArtefactRow {
+    artefact_id: String,
+    symbol_id: String,
+    language: String,
+}
+
+#[cfg(test)]
+#[derive(Debug, Clone)]
+struct FunctionArtefact {
+    name: String,
+    start_line: i32,
+    end_line: i32,
+    start_byte: i32,
+    end_byte: i32,
+    signature: String,
+}
+
+#[derive(Debug, Clone)]
+struct JsTsArtefact {
+    canonical_kind: String,
+    language_kind: String,
+    name: String,
+    symbol_fqn: String,
+    parent_symbol_fqn: Option<String>,
+    start_line: i32,
+    end_line: i32,
+    start_byte: i32,
+    end_byte: i32,
+    signature: String,
+}
+
+#[derive(Debug, Clone)]
+struct JsTsDependencyEdge {
+    edge_kind: String,
+    from_symbol_fqn: String,
+    to_target_symbol_fqn: Option<String>,
+    to_symbol_ref: Option<String>,
+    start_line: Option<i32>,
+    end_line: Option<i32>,
+    metadata: Value,
+}
+
+#[derive(Debug)]
+struct RustUseExportEntry {
+    path: String,
+    export_name: String,
+}
