+++ bitloops_cli/src/engine/mod.rs
@@ -1,5 +1,6 @@
 pub mod agent;
 pub mod db_status;
+pub mod devql;
 pub mod git_operations;
 pub mod history;
 pub mod hooks;
